Переделана система перехвата исключений. Теперь исключения обрабатываются в диспетчере.

+++ Yanitta/App.xaml.cs
@@ -17,6 +17,21 @@ namespace Yanitta
             ProcessList = new ProcessList();
         }
 
+        public App()
+            : base()
+        {
+            this.Dispatcher.UnhandledException += (o, e) =>
+            {
+                if (e.Exception is YanittaException)
+                {
+                    MessageBox.Show(e.Exception.Message,
+                        "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    Console.WriteLine(e.Exception.Message);
+                    e.Handled = true;
+                }
+            };
+        }
+
         protected override void OnStartup(StartupEventArgs e)
         {
             if (e.Args.Length > 0 && e.Args[0] == "/e")
@@ -30,38 +45,48 @@ namespace Yanitta
             var fileName = Settings.Default.ProfilesFileName;
             Console.WriteLine("Yanitta startup!...");
 
-            try
+            // Если файл профиля отсутствует, создаем новый (пустой) и запускаем редактор профилей.
+            if (!File.Exists(fileName))
             {
-                ProfileDb.Instance = XmlManager.Load<ProfileDb>(fileName);
-
-                //Если загрузка прошла удачно - сделаем резервную копию.
-                if (File.Exists(fileName))
-                    File.Copy(fileName, fileName + ".bak", true);
-
-                Console.WriteLine("База успешно загружена, резервная копия создана.");
+                ProfileDb.Instance = new ProfileDb();
+                File.WriteAllText(fileName, "");
+                StartupUri = new Uri("Windows/WinProfileEditor.xaml", UriKind.Relative);
             }
-            catch (InvalidOperationException ex)
+            else
             {
-                Console.WriteLine("Ошибка загрузки базы данных, возможно база повреждена.");
-                Console.WriteLine(ex.Message);
-                // Попытка загрузить резервную базу.
-                if (File.Exists(fileName + ".bak"))
-                {
-                    Console.WriteLine("Найден файл резервной копии...");
-                    File.Copy(fileName + ".bak", fileName, true);
-                }
-                else
-                    throw new FileNotFoundException("Файл резервной копии не найден.", fileName + ".bak");
-
                 try
                 {
-                    Console.WriteLine("Попытка загрузить базу данных из резервной копии.");
                     ProfileDb.Instance = XmlManager.Load<ProfileDb>(fileName);
-                    Console.WriteLine("База из резервной копии успешно загружена.");
+
+                    //Если загрузка прошла удачно - сделаем резервную копию.
+                    if (File.Exists(fileName))
+                        File.Copy(fileName, fileName + ".bak", true);
+
+                    Console.WriteLine("База успешно загружена, резервная копия создана.");
                 }
-                catch (Exception ex_inner)
+                catch (InvalidOperationException ex)
                 {
-                    throw new YanittaException("Ошибка загрузки базы из резервной копии.\r\n" + ex_inner.Message);
+                    Console.WriteLine("Ошибка загрузки базы данных, возможно база повреждена.");
+                    Console.WriteLine(ex.Message);
+                    // Попытка загрузить резервную базу.
+                    if (File.Exists(fileName + ".bak"))
+                    {
+                        Console.WriteLine("Найден файл резервной копии...");
+                        File.Copy(fileName + ".bak", fileName, true);
+                    }
+                    else
+                        throw new FileNotFoundException("Файл резервной копии не найден.", fileName + ".bak");
+
+                    try
+                    {
+                        Console.WriteLine("Попытка загрузить базу данных из резервной копии.");
+                        ProfileDb.Instance = XmlManager.Load<ProfileDb>(fileName);
+                        Console.WriteLine("База из резервной копии успешно загружена.");
+                    }
+                    catch (Exception ex_inner)
+                    {
+                        throw new YanittaException("Ошибка загрузки базы из резервной копии.\r\n" + ex_inner.Message);
+                    }
                 }
             }
